Revert "rename RxWebSocket -> WalletLinkWebSocket"
This reverts commit c3ae669ff2f7c47cab79bf5ae425c6b54963c58f.
