+++ packages/wallet-sdk/src/connection/DiagnosticLogger.ts
@@ -2,8 +2,8 @@
 
 import { Web3RequestMessage } from '../relay/Web3RequestMessage';
 import { Web3ResponseMessage } from '../relay/Web3ResponseMessage';
+import { ConnectionState } from './RxWebSocket';
 import { ServerMessage, ServerMessageIsLinkedOK } from './ServerMessage';
-import { ConnectionState } from './WalletLinkWebSocket';
 
 export type LogProperties = {
   addresses_length?: number; // number of eth addresses
